server/worker/join_credits/requirements.txt
aio-pika==9.5.5
python-dotenv==1.1.0
